requirements.txt
alembic==1.7.7
attrs==21.4.0
click==8.1.2
coverage==6.3.2
Flask==2.1.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
pytest-flask==1.2.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
SQLAlchemy-Utils==0.38.2
tomli==2.0.1
Werkzeug==2.1.1
zipp==3.8.0
